+++ keanu-project/src/main/java/io/improbable/keanu/tensor/TensorShape.java
@@ -6,6 +6,7 @@ import org.apache.commons.lang3.ArrayUtils;
 
 import java.util.Arrays;
 
+
 public class TensorShape {
 
     private long[] shape;